requirements.txt
django
django-environs
psycopg2-binary
redis
celery
djangorestframework
pillow
drf-spectacular
requests